game/requirements.txt
pandas
numpy
openai
Pillow
tensorflow
opencv-python
mediapipe
requests